requirements.txt
streamlit
numpy
pandas
scikit-learn
Pillow
xgboost


#PIL は pillow とかく